cleaned a bit
